+++ docs/source/history.rst
@@ -13,6 +13,12 @@ V1.7.0
 ADC/DAC:
     * Fixed bug. When GP1, 2 and 3 are all in ADC mode and ADC reference is VDD. If just after reset VRM reference is selected, ADC stops working.
 
+IOC:
+    * Added new section about Interrupt On Change.
+    * Renamed function *wake_up_config* to :func:`IOC_config`.
+    * Added :func:`IOC_read` to read Interrupt On Change flag.
+    * Added :func:`IOC_clear` to clear Interrupt On Change flag.
+
 I2C:
     * Removed deprecated functions *I2C_is_idle* and *I2C_cancel*.